+++ src/y2021/d05/parser.rs
@@ -13,7 +13,7 @@ pub(super) fn coordinate(s: &str) -> IResult<&str, Coordinate> {
     let (s, x) = i64(s)?;
     let (s, _) = tag(",")(s)?;
     let (s, y) = i64(s)?;
-    Ok((s, Coordinate::new(x, y)))
+    Ok((s, Coordinate::new(x as isize, y as isize)))
 }
 
 pub(super) fn line(s: &str) -> IResult<&str, Line> {
