+++ src/y2021/d17/part2.rs
@@ -29,8 +29,8 @@
 */
 
 #[macros::problem]
-fn problem(input: &super::TargetArea) -> isize {
-    input.range().count() as isize
+fn problem(input: &super::TargetArea) -> usize {
+    input.range().count()
 }
 
 #[cfg(test)]
